# feat: reduce internal resolution and scale