## PR type (check all applicable)

- [x] `   feat   ` :sparkles: features
- [ ] `   fix    ` :bug: bugfixes
- [x] `  chore   ` :ticket: chores
- [x] `refactor` :package: code refactoring
- [ ] `  style   ` :gem: style
- [ ] `   docs   ` :book: documentation changes
- [ ] `   perf   ` :rocket: performance improvements
- [ ] `   test   ` :rotating_light: tests
- [ ] `  build   ` :construction_worker: build
- [ ] `    ci    ` :robot: continuous integration
- [ ] `  revert  ` :back: reverts

## PR description

This PR:
- Decreases the resolution of the game to `640x360` (2x smaller).
- Decreases tile size to 32px.

## Related Tickets

- related issue #
- closes #

## Instructions, Screenshots

Peformance improvements:

| Revision                            | Main Menu  | Game Map   |
|-------------------------------------|------------|------------|
| Before (1280x720 native)            | 498.26 FPS | 296.80 FPS |
| After (640x360, scaled to 1280x720) | 572.10 FPS | 430.14 FPS |
| Performance improvement             | +14.92%    | +44.92%    |

Graphical comparison:

Scaled:
![scaling_640x360_new](https://github.com/pufereq/simulat/assets/94570596/13aeba89-0384-4fc2-955d-d3dcc0243257)

Native:
![scaling_old_0.13.0](https://github.com/pufereq/simulat/assets/94570596/3db21429-dbf0-480f-b544-fb81c06cdf7f)